move dataclasses to transport2030 energy_demand
