+++ src/climatevision/generator/transport2030/energy_demand/road.py
@@ -2,11 +2,12 @@
 
 from dataclasses import dataclass, asdict
 
-from ..inputs import Inputs
-from ..utils import div
-from ..transport2018.t18 import T18
+from ...inputs import Inputs
+from ...utils import div
+from ...transport2018.t18 import T18
 
-from .transport import Transport
+from ..transport import Transport
+
 from .investmentaction import InvestmentAction, RoadInvestmentAction
