+++ src/climatevision/generator/transport2030/t30.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass, asdict
 
 from ..common.energy import Energy
 
-from .investmentaction import InvestmentAction
-from .air import Air
-from .road import (
+from .energy_demand import (
+    InvestmentAction,
+    Air,
     Road,
     RoadBus,
     RoadCar,
@@ -16,10 +16,19 @@ from .road import (
     RoadPeople,
     RoadSum,
     RoadInvestmentAction,
+    Rail,
+    RailGoods,
+    RailPeople,
+    RailPeopleMetroActionInfra,
+    RailPeopleSum,
+    Ship,
+    ShipDomestic,
+    ShipInternational,
+    ShipDomesticActionInfra,
+    Other,
+    OtherCycle,
+    OtherFoot,
 )
-from .rail import Rail, RailGoods, RailPeople, RailPeopleMetroActionInfra, RailPeopleSum
-from .ship import Ship, ShipDomestic, ShipInternational, ShipDomesticActionInfra
-from .other import Other, OtherCycle, OtherFoot
 from .transport import Transport
 from .dataclasses import GPlanning, G, T
